server: move constants into lib
