+++ server/handlers/error.ts
@@ -4,7 +4,7 @@ import fs from "fs";
 import * as Sentry from "@sentry/node";
 import express from "express";
 
-import { TEMPLATE_DIR } from "../constants";
+import { TEMPLATE_DIR } from "../lib/constants";
 
 const errorPage = fs
   .readFileSync(path.join(TEMPLATE_DIR, "error-message.html"))
